Implement support for InactiveDestroy
When InactiveDestroy is enabled on $dbh then DBI driver must ensure that
when $dbh refcount goes to zero, it does not shutdown connection with the
server, but at the same time needs to ensure that driver does not leak any
memory or state related to $dbh structures.

Explicit call to $dbh->disconnect() or $drh->disconnect_all() is not
affected by the $dbh->{InactiveDestroy} attribute. When doing explicit call
to $dbh->disconnect() or $drh->disconnect_all(), it always has to send
disconnect command to the server.

Note that DBI calls $drh->disconnect_all() in its END block and this call
has to be affected by InactiveDestroy attribute. DBI sets $DBI::PERL_ENDING
to 1 to allow drivers detecting between when is DBI's END block running.

As MariaDB client library and neither MySQL client library does not provide
required API for implementing InactiveDestroy, use some kind of hack.

There is only function mysql_close() which is doing two things: Shutdown
connection between client and server, and release memory associated with
mysql structures.

So before calling mysql_close(), change system socket used by the mysql
library by some other socket on which can be freely send MySQL command for
quitting connection. For UNIX systems, it is replaced by the new socket of
AF_UNIX/SOCK_STREAM type by dup2() syscall which replaced file descriptor
to the existing value. So mysql_close() would not be able to send QUIT
command to the original server, but rather to the new unconnected socket.
On Windows systems, only the original socket is closed without doing any
replacement as it looks like that Windows does not provide kernel syscall
which can create a new socket with explicit SOCKET identifier.

When InactiveDestroy is enabled on $sth then same happens but for prepared
statement on the server. DBI driver must ensure that it is not removed on
the server. For MariaDB and MySQL client libraries, this is achieved by
setting stmt->state to zero which prevents mysql_stmt_close() to send
destroy command to the server.